+++ src/utils/deleteFile.js
@@ -1,6 +1,9 @@
 import fs from 'fs';
 
 function deleteFile(pathList) {
+  if (!pathList) {
+    return false;
+  }
   pathList.forEach((element) => fs.unlinkSync(element.path));
 }
